[DOP-24163] Add handleCallback to keycloakProvider
