+++ src/auth/utils/getAuthProvider.tsx
@@ -4,7 +4,7 @@ import { Login, keycloakLoginForm } from "@/components/login";
 
 // replaced by docker image entrypoint
 /* eslint-disable prefer-const */
-let AUTH_PROVIDER = "dummyAuthProvider";
+let AUTH_PROVIDER = "keycloakAuthProvider";
 
 const getAuthProvider = () => {
     switch (AUTH_PROVIDER) {
