+++ src/components/login/index.ts
@@ -1,4 +1,4 @@
-import { KeycloakAuthCallback, keycloakLoginForm } from "./keycloak";
+import { keycloakLoginForm } from "./keycloak";
 import Login from "./Login";
 
-export { KeycloakAuthCallback, Login, keycloakLoginForm };
+export { Login, keycloakLoginForm };
