+++ src/components/login/keycloak/index.ts
@@ -1,4 +1,3 @@
 import keycloakLoginForm from "./keycloak_login";
-import KeycloakAuthCallback from "./callback";
 
-export { keycloakLoginForm, KeycloakAuthCallback };
+export { keycloakLoginForm };
